requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
breadability==0.1.20
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==4.2.18
docopt==0.6.2
filelock==3.16.1
flatbuffers==24.12.23
fsspec==2024.12.0
gast==0.6.0
gensim==4.3.3
gensim_sum_ext==0.1.2
google-pasta==0.2.0
grpcio==1.69.0
h5py==3.12.1
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
joblib==1.4.2
keras==3.8.0
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
nltk==3.9.1
numpy==1.26.4
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pillow==11.1.0
pip-chill==1.0.3
protobuf==5.29.3
pycorenlp==0.3.0
pycountry==24.6.1
Pygments==2.19.1
PyMuPDF==1.25.1
pytesseract==0.3.13
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scipy==1.13.1
six==1.17.0
smart-open==7.1.0
sqlparse==0.5.3
sumy==0.11.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
